pyproject.toml
[project]
name = "orbital-junkyard"
version = "0.1.0"
description = "Space Debris Intelligence Platform powered by Databricks"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pyspark>=3.5.0",
    "delta-spark>=3.1.0",
    "schedule>=1.2.0",
]
